// src/app/Models/GenStatementMdl.ts
import { nullSafeIsEquivalent } from "@angular/compiler/src/output/output_ast";

export class GenstatementMdl{
    // RentalAgreementID: number;
    // PaymentMonth: string;
    // // File: File;
    // fileName: string = "";
    // fileSize: number = 0;
    // fileType: string = "";
    // lastModifiedTime: number = 0;
    // lastModifiedDate: Date = null;
    // fileAsBase64: string = "";

    BuildingID: number = 0;
    UnitID: number = 0;
    StartDate: Date = null;
    EndDate: Date = null;
}